--- Backend/Helpers/BankingServiceSettings.cs
@@ -1,10 +1,0 @@
-namespace Backend.Helpers;
-
-public class BankingServiceSettings
-{
-    public string? RetailKey { get; set; }
-    public string? CommercialKey { get; set; }
-    public string? CompanyAccount { get; set; }
-    public string? RetailEndpoint { get; set; }
-    public string? CommercialEndpoint { get; set; }
-}
