--- Backend/Models/DeregisterElectronics.cs
@@ -1,7 +1,0 @@
-
-namespace Backend.Models;
-
-public class DeregisterElectronics(long personaId)
-{
-  public long PersonaId { get; set; } = personaId;
-}
