--- Backend/Models/TaxNotification.cs
@@ -1,10 +1,0 @@
-
-
-namespace Backend.Models;
-
-public class TaxNotification(double amount, string bankAccount)
-{
-  public double Amount { get; set; } = amount;
-  
-  public string BankAccount { get; set; } = bankAccount;
-}
