--- Backend/Models/TimeNotification.cs
@@ -1,6 +1,0 @@
-namespace Backend.Models;
-
-public class TimeNotification(string currentTime)
-{
-  public string CurrentTime { get; set;} = currentTime;
-}
